+++ open-xchange-appsuite-public-sector/src/io.ox.public-sector/settings.js
@@ -1,3 +1,16 @@
 import { Settings } from '$/io.ox/core/settings'
 
+export const defaults = {
+  autonotify: {
+    subject: {
+      default: 'There are new items in your mailbox',
+      de_DE: 'Es liegen neue Nachrichten in Ihrem Postfach'
+    },
+    body: {
+      default: 'There are new items in your mailbox',
+      de_DE: 'Es liegen neue Nachrichten in Ihrem Postfach'
+    }
+  }
+}
+
 export const settings = new Settings('io.ox.public-sector', () => ({}))